src/requirements.txt
fastapi==0.92.0
python-dotenv==1.0.0
SQLAlchemy==1.3.20
